+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased] - 2017-02-07
+### Changed
+- Implement `OptionRepository::findAll()` and fix `OptionsStore::getAll()`. Props @gmazzap.
+
 ## [0.1.5] - 2017-02-07
 ### Fixed
 - Removed `void` return type hint from `IdentityMap::purge()`.